Move other model utils into /utils/models
