+++ src/Hedwig/ClientApp/src/containers/Reports/ReportDetail/UtilizationTable.tsx
@@ -4,8 +4,7 @@ import { CdcReport, Enrollment, Age, FundingSource, FundingTime, Region } from '
 import idx from 'idx';
 import moment from 'moment';
 import { CdcRates } from './CdcRates';
-import { prettyAge } from '../../../utils/models';
-import { prettyFundingTime } from '../../../utils/fundingTimeUtils';
+import { prettyAge, prettyFundingTime } from '../../../utils/models';
 import currencyFormatter from '../../../utils/currencyFormatter';
 import cartesianProduct from '../../../utils/cartesianProduct';
