+++ src/Hedwig/ClientApp/src/utils/models/index.ts
@@ -5,3 +5,5 @@ export * from './familyDetermination';
 export * from './funding';
 export * from './enrollment';
 export * from './ageGroup';
+export * from './fundingTime';
+export * from './reportingPeriod';
